+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2023-08-12
+### Added
+- 'logsDirectory' option
+- check if the logs directory exists before initialising plugin
+
+
 ## [1.0.1] - 2023-02-19
 ### Added
 - README.md